+++ CHANGELOG.md
@@ -1,9 +1,25 @@
 # ChangeLog
 
+## 2025-9-20
+
+### Added
+  - funcion verficar_fichas_sacadas_15 en clase Boaed
+  - test test_verficar_fichas_sacadas_15_negro en test
+  - test test_verficar_fichas_sacadas_15_blanco en test
+
+## 2025-9-20
+
+### Added
+  - clase player con core/player.py
+  - funcion __init__ en clase Player
+  - funciones para obtener atributos de clase Player
+  - funcion crear jugador en clase BlackgammonGame
+  - test test_crear_jugador en test
+
 ## 2025-9-14
 
 ### Added
-  - Funcio verificar_cambio_turno en BackgammonGame
+  - Funcion verificar_cambio_turno en BackgammonGame
   - test  test_varificar_cambio_turno_cambia en test
   - test  test_varificar_cambio_turno_no_cambia test